Fix a rentals retrieving after restart bug
