+++ books.json
@@ -1,1 +1,1 @@
-[{"title":"DD","author":"RE"},{"title":"VV","author":"YK"}]
+["{\"json_class\":\"Book\",\"a\":[\"ewe\",\"au\"]}"]
