--- create_student.rb
@@ -1,22 +1,0 @@
-require_relative 'student'
-require_relative 'person'
-require_relative 'nameable'
-def create_a_student
-  print 'Age: '
-  age = gets.chomp
-  print 'Name: '
-  name = gets.chomp
-  print 'Has parent permission? [Y/N]: '
-  permission = gets.chomp
-  case permission
-  when 'y', 'Y'
-    @persons.push(Student.new(true, age, name))
-    puts 'Student created successfully'
-  when 'n', 'N'
-    @persons.push(Student.new(false, age, name))
-    puts 'Student created successfully'
-  else
-    puts 'Invalid permission (Enter Y/N)'
-    library_menu
-  end
-end
