+++ people.json
@@ -1,1 +1,1 @@
-[{"type":"Student","name":"Sam","id":74,"age":24,"classroom":"FR","parent_permission":"true"},{"type":"Teacher","name":"DD","id":236,"age":43,"specialization":"HYD"}]
+["{\"json_class\":\"Student\",\"a\":[11,\"math\",\"ee\",\"true\",284]}"]
