+++ rental.rb
@@ -8,4 +8,15 @@ class Rental
     @person = person
     person.rentals << self
   end
+
+  def to_json(*args)
+    {
+      JSON.create_id => self.class.name,
+      'a' => [@date, @book, @person]
+    }.to_json(*args)
+  end
+
+  def self.json_create(object)
+    new(*object['a'])
+  end
 end
