+++ rentals.json
@@ -1,1 +1,1 @@
-[{"date":"12/01/2023","person_id":74,"book_title":"DD"},{"date":"09/12/23","person_id":74,"book_title":"VV"}]
+[{"date":"2020-12-12","person_id":284,"book_title":"ewe"}]
